a mistake

+++ التجربة الثالثة/index.ino
@@ -7,8 +7,8 @@ void setup() {
 }
 
 void loop() {
-    int ADCin = analogRead(A0);
-    int ADCout = map(ADCin ,0,1023,0,255);
+    float ADCin = analogRead(A0);
+    float ADCout = map(ADCin ,0,1023,0,255);
     if (ADCin > 650)
         analogWrite(led, ADCout);
 }
